chore: Mapper 구현
기존의 BaseResponse 를 ApiResponse 로 변경
