+++ client/data/src/main/java/com/everyone/data/mapper/Mapper.kt
@@ -1,0 +1,8 @@
+package com.everyone.data.mapper
+
+import com.everyone.data.base.BaseResponse
+import com.everyone.domain.model.base.BaseModel
+
+interface Mapper<in R : BaseResponse, out D : BaseModel> {
+    fun R.toDomainModel(): D
+}
