+++ client/domain/src/main/java/com/everyone/domain/model/base/BaseModel.kt
@@ -1,0 +1,5 @@
+package com.everyone.domain.model.base
+
+import android.os.Parcelable
+
+interface BaseModel : Parcelable
